Set up koin modules for platform-dependent stuff
