+++ data/src/commonMain/kotlin/fobo66/valiutchik/core/di/ThirdPartyModule.kt
@@ -16,31 +16,6 @@
 
 package fobo66.valiutchik.core.di
 
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.PreferenceDataStoreFactory
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.preferencesDataStoreFile
-import androidx.room.Room
-import androidx.sqlite.driver.AndroidSQLiteDriver
-import fobo66.valiutchik.core.db.CurrencyRatesDatabase
-import org.koin.android.ext.koin.androidContext
-import org.koin.dsl.module
+import org.koin.core.module.Module
 
-val thirdPartyModule = module {
-
-    single<CurrencyRatesDatabase> {
-        val dbFile = androidContext().getDatabasePath("currency-rates")
-        Room.databaseBuilder<CurrencyRatesDatabase>(
-            context = androidContext(),
-            name = dbFile.absolutePath
-        )
-            .setDriver(AndroidSQLiteDriver())
-            .build()
-    }
-
-    single<DataStore<Preferences>> {
-        PreferenceDataStoreFactory.create {
-            androidContext().preferencesDataStoreFile("valiutchik-prefs")
-        }
-    }
-}
+expect val thirdPartyModule: Module
